Rename sonata_consts in sonata_constants
